requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
gitdb==4.0.12
GitPython==3.1.44
python-dotenv==1.1.0
smmap==5.0.2
sqlparse==0.5.3
tzdata==2025.2
